replace fill order details action with convert payment
