--- Action/FillOrderDetailsAction.php
@@ -1,55 +1,0 @@
-<?php
-
-namespace Ledjin\Sagepay\Action;
-
-use Payum\Core\Action\ActionInterface;
-use Payum\Core\Exception\RequestNotSupportedException;
-use Payum\Core\Request\FillOrderDetails;
-use Payum\Core\Security\GenericTokenFactoryInterface;
-
-class FillOrderDetailsAction implements ActionInterface
-{
-
-    /**
-     * @var GenericTokenFactoryInterface
-     */
-    protected $tokenFactory;
-
-    /**
-     * @param GenericTokenFactoryInterface $tokenFactory
-     */
-    public function __construct(GenericTokenFactoryInterface $tokenFactory = null)
-    {
-        $this->tokenFactory = $tokenFactory;
-    }
-
-    /**
-     * {@inheritDoc}
-     *
-     * @param FillOrderDetails $request
-     */
-    public function execute($request)
-    {
-        RequestNotSupportedException::assertSupports($this, $request);
-
-        $order = $request->getOrder();
-
-        $details = $order->getDetails();
-
-        $details['VendorTxCode'] = $order->getNumber();
-        $details['Amount'] = $order->getTotalAmount() / 100;
-        $details['Currency'] = $order->getCurrencyCode();
-        $details['Description'] = $order->getDescription();
-        $details['CustomerEMail'] = $order->getClientEmail();
-
-        $order->setDetails($details);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function supports($request)
-    {
-        return $request instanceof FillOrderDetails;
-    }
-}
